% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{preconditions}
\alias{preconditions}
\title{Check inputs against known limitations}
\usage{
preconditions(x)
}
\arguments{
\item{x}{a scalar (length 1) numeric vector to be compared to a
random number.}
}
\value{
A scalar numeric vector. If the input is greater than a
    random normal deviate, the return value is the random normal
    deviate. Otherwise, the return value is the orignal input.
}
\description{
Checking inputs allows code to \sQuote{fail early} with clear error
messages, rather tha failing cryptically later.
}
\details{
Update the code below to check that the variable used in the
\code{if} is scalar, to avoid a mysterious message \dQuote{the
condition has length > 1 and only the first element will be
used}. Use \code{stopifnot()} as a convenient way to test the pre-
conditions.

Find examples on the Bioconductor support site where this error has
confused users.
}
\examples{
preconditions(.1)
}

