% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{sequences}
\alias{sequences}
\title{Generate sequences correctly}
\usage{
sequences(x)
}
\arguments{
\item{x}{a numeric vector of any length}
}
\value{
A character vector of the same length as \code{x}
}
\description{
A very common problem is the use of 1:n to generate a sequence. The
following illustrates the problem; what's the solution?
}
\examples{
sequences(1:5)         # success
sequences(integer(0))  # failure!
}

