% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficient.R
\name{vectorize}
\alias{vectorize}
\title{Vectorize!}
\usage{
vectorize(x)
}
\arguments{
\item{x}{a numeric vector of non-negative numbers of any length}
}
\value{
A numeric vector of \code{length(x)}, where each element is
    the square root of x
}
\description{
People often confuse iteration and vectorization.
}
\details{
The implementation below is an iteration. Vectorize it.

Show that vectorization influences performance using the
microbenchmark package.

Find one StackOverflow post where the question or answer confuses
iteration and vectorization.
}
\examples{
vectorize(1:10)
}

